Refactoring Repository Interface
