+++ app/Contracts/Services/Auth/AuthServiceInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Contracts\Services;
+namespace App\Contracts\Services\Auth;
 
 use App\Models\User;
