+++ app/Repositories/Item/ItemRepository.php
@@ -2,6 +2,7 @@
 
 namespace App\Repositories\Item;
 
+use App\Contracts\Repositories\OrderItem\ItemRepositoryInterface;
 use App\Models\{OrderItem};
 
 class ItemRepository implements ItemRepositoryInterface
