+++ app/Repositories/User/UserRepository.php
@@ -2,6 +2,7 @@
 
 namespace App\Repositories\User;
 
+use App\Contracts\Repositories\User\UserRepositoryInterface;
 use App\Models\User;
 
 class UserRepository implements UserRepositoryInterface
