+++ app/Services/Category/UpdateCategoryService.php
@@ -3,7 +3,7 @@
 namespace App\Services\Category;
 
 use App\Models\Category;
-use App\Repositories\Category\CategoryRepositoryInterface;
+use App\Contracts\Repositories\Category\CategoryRepositoryInterface;
 use Illuminate\Support\Str;
 
 class UpdateCategoryService
