+++ app/Services/User/CreateUserService.php
@@ -3,7 +3,7 @@
 namespace App\Services\User;
 
 use App\Models\User;
-use App\Repositories\User\UserRepositoryInterface;
+use App\Contracts\Repositories\User\UserRepositoryInterface;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Str;
